Adding Script to Combine MP3 Post Download to M4B
This "should" work. It is a new script that will be designed for audiobooks. I included ffmpeg in the build itself, verse a docker mod.  The goal of this script is to recreate what calibre-server does for metadata, Assuming this PR is approved (and I may need some help) I want to add tone and metadata. This is a good first step by combining multiple mp3s and even m4bs into one clean file.